Moves auth code to appropriate folders
