+++ src/main/java/ca/uhn/fhir/jpa/starter/common/StarterJpaConfig.java
@@ -40,13 +40,13 @@ import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
 import ca.uhn.fhir.jpa.search.IStaleSearchDeletingSvc;
 import ca.uhn.fhir.jpa.search.StaleSearchDeletingSvcImpl;
 import ca.uhn.fhir.jpa.starter.AppProperties;
-import ca.uhn.fhir.jpa.starter.CustomAuthorizationInterceptor;
-import ca.uhn.fhir.jpa.starter.CustomConsentService;
-import ca.uhn.fhir.jpa.starter.CustomSearchNarrowingInterceptor;
 import ca.uhn.fhir.jpa.starter.annotations.OnCorsPresent;
 import ca.uhn.fhir.jpa.starter.annotations.OnImplementationGuidesPresent;
 import ca.uhn.fhir.jpa.starter.common.validation.IRepositoryValidationInterceptorFactory;
 import ca.uhn.fhir.jpa.starter.interceptor.CapabilityStatementCustomizer;
+import ca.uhn.fhir.jpa.starter.interceptor.CustomAuthorizationInterceptor;
+import ca.uhn.fhir.jpa.starter.interceptor.CustomConsentService;
+import ca.uhn.fhir.jpa.starter.interceptor.CustomSearchNarrowingInterceptor;
 import ca.uhn.fhir.jpa.starter.ips.IpsConfigCondition;
 import ca.uhn.fhir.jpa.starter.util.EnvironmentHelper;
 import ca.uhn.fhir.jpa.subscription.util.SubscriptionDebugLogInterceptor;
